Created nav-bar component with styling. Added global styling for anchor tags.
