+++ frontend-react/src/App.js
@@ -1,7 +1,9 @@
 import './styles/global.css';
 
 import { Routes, Route } from "react-router-dom";
-import Nav from "./components/Nav.jsx";
+import Nav from "./components/nav/Nav.jsx";
+
+
 
 function App() {
     return (
