+++ frontend-react/src/components/nav/Nav.css
@@ -1,0 +1,111 @@
+.menubar{
+    background-image: linear-gradient( rgba(12, 11, 11, 0.45), rgb(236, 113, 111)), url("../../images/longdoctor.jpeg");
+    background-repeat: no-repeat;
+    background-attachment: fixed;
+    background-size: cover;
+    height: 60vh;
+    display: flex;
+    flex-direction: column;
+}
+
+.nav-toggler {
+    display: none;
+}
+
+.header-container {
+    display: flex;
+    padding: 3.2rem 3.2rem;
+    justify-content: space-between;
+    font-size: 2rem;
+}
+
+.menubar-links{
+    display: flex;
+    align-items: center;
+    justify-content: space-evenly;
+    gap: 3rem;
+}
+
+.brand-image img {
+    height: 50px;
+    width: auto;
+}
+
+.container-text {
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    text-align: center;
+    align-items: center;
+    flex-wrap: wrap;
+    margin-top: 10rem;
+    gap: 20px;
+    font-size: 3.5rem;
+    color: rgb(255, 255, 255);
+}
+
+.signup-btn button {
+    background-color: #ec361e;
+    border-color: #ffffff;
+    border-style: solid;
+    border-radius: 0.3rem;
+    color: white;
+    padding: 1.5rem 3.2rem;
+    font-size: 1.6rem;
+    margin-top: 4rem;
+}
+
+.signup-btn button:hover {
+    background-color: #8a2315;
+}
+
+.nav-toggler {
+    display: none;
+}
+
+.nav-toggler div {
+    width: 2.5rem;
+    height: 0.2rem;
+    margin: 0.4rem;
+    background: rgb(204, 204, 204);
+    transition: 0.3s ease-in;
+}
+
+@media (max-width: 600px) {
+
+    .nav-toggler {
+        display: flex;
+        flex-direction: column;
+        justify-content: center;
+        cursor: pointer;
+    }
+    .menubar-links {
+        position: fixed;
+        top: 15vh;
+        right: 0;
+        height: 85vh;
+        width: 100%;
+        background: linear-gradient( rgb(105, 103, 103), rgb(140, 57, 54));
+        flex-direction: column;
+        transform: translateX(100%);
+        transition: 0.5s ease-in;
+    }
+}
+
+
+/* Active Class */
+.nav-active {
+    transform: translateX(0%);
+}
+
+/* Toggle Icon Animation */
+
+.toggle .line1 {
+    transform: rotate(-45deg) translate(-8.5px, 5px);
+}
+.toggle .line2 {
+    opacity: 0;
+}
+.toggle .line3 {
+    transform: rotate(45deg) translate(-10px, -5px);
+}
