+++ frontend-react/src/styles/global.css
@@ -40,6 +40,7 @@ h1, h2, h3, h4, h5 {
 
 a {
     text-decoration: none;
+    color: white;
 }
 
 li{
